pyproject.toml
[project]
name = "doc-intelligence-app"
version = "0.1.0"
description = "Streamlit application for document intelligence with Databricks integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "databricks-agents",
    "databricks-sdk>=0.18.0",
    "databricks-connect==16.3.2",
    "databricks-langchain==0.6.0",
    "mlflow>=3.1",
    "langgraph",
    "langchain-postgres",
    "langchain-community",
    "langchain-text-splitters",
    "langgraph-checkpoint-postgres",
    "pgvector",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.28.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "aiofiles",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = [
    "src/",
]
testpaths = [
    "tests",
]